fix(graph): address R1 final-review findings (I1 flat owned-context, M2/M4/M5)
- I1: resources_in_team_scope owned-context branch self-gates on direct membership
  (matches the model's flat treatment; no reliance on the trailing intersection).
- M2: e2e asserts squad-a zone resource_count == 1.
- M4: TeamZone.resource_count is i32 (count(*)::int) -> TS number.
- M5: comment the defensive team-fetch ok_or.
Full cargo make check + SQL/e2e/type tests green.

+++ crates/temper-core/src/types/graph_scope.rs
@@ -28,7 +28,7 @@ pub struct TeamZone {
     pub id: Uuid,
     pub slug: String,
     pub name: String,
-    pub resource_count: i64,
+    pub resource_count: i32,
 }
 
 /// The team-scoped navigation frame for the graph view: the scope team, its reachable
